+++ src/server/js/functions.js
@@ -9,6 +9,7 @@ import { Accounts } from 'meteor/accounts-base';
 AccountsTools.server = {
     /*
      * @param {String} the searched email address
+     * @param {Object} options an optional dictionary of fields to return or exclude
      * @returns {Promise} which eventually resolves to the user document, or null
      * 
      *  As a reminder, see https://v3-docs.meteor.com/api/accounts.html#Meteor-users
@@ -47,6 +48,33 @@ AccountsTools.server = {
                     doc = AccountsTools.cleanupUserDocument( doc );
                     if( AccountsTools.opts().verbosity() & AccountsTools.C.Verbose.SERVERDB ){
                         console.log( 'pwix:accounts-tools byId('+id+')', doc );
+                    }
+                    return doc;
+                });
+        } else {
+            // either a code error or a user try to bypass our checks
+            throw new Meteor.Error( 'arg', 'incorrect argument' );
+        }
+    },
+
+    /*
+     * @param {String} the searched username
+     * @param {Object} options an optional dictionary of fields to return or exclude
+     * @returns {Promise} which eventually resolves to the user document, or null
+     * 
+     *  As a reminder, see https://v3-docs.meteor.com/api/accounts.html#Accounts-findUserByUsername
+     *  Each username can only belong to one user
+     *  In other words, a username can be considered as a user identiier in Meteor ecosystems
+     */
+    async byUsername( username, options={} ){
+        check( username, String );
+        check( options, Object );
+        if( username && _.isString( username )){
+            return Accounts.findUserByUsername( username, options )
+                .then(( doc ) => {
+                    doc = AccountsTools.cleanupUserDocument( doc );
+                    if( AccountsTools.opts().verbosity() & AccountsTools.C.Verbose.SERVERDB ){
+                        console.log( 'pwix:accounts-tools byUsername('+username+')', doc );
                     }
                     return doc;
                 });